requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlmodel==0.0.24
starlette==0.47.2
psycopg[binary]==3.1.19        
python-jose==3.5.0
cryptography==45.0.6          
httpx==0.28.1
python-dotenv==1.1.1
PyJWT>=2.9,<3